+++ vitest.config.js
@@ -2,6 +2,7 @@ import { defineConfig, configDefaults } from 'vitest/config';
 
 export default defineConfig({
   test: {
+    cacheDir: './.cache/vitest',
     coverage: {
       provider: 'v8',
       reporter: ['text', 'json', 'lcov', 'html'],
@@ -11,10 +12,10 @@ export default defineConfig({
         'commitlint.config.js',
         'eslint.ignores.js',
         'vitest.workspace.js',
-        './packages/observables/src/index.js',
-        './packages/operators/src/index.js'
+        './packages/mocks/',
+        './packages/*/src/index.js'
       ]
     },
-    include: ['**/packages/**/*.test.js']
+    include: ['./packages/**/*.test.js']
   }
 });